refactor: 순위의 결과를 가지고 있는 컬렉션을 일급컬렉션으로 포장

+++ src/main/java/lotto/domain/LottoRank.java
@@ -42,7 +42,7 @@ public enum LottoRank {
         return this.matchCount == matchCount;
     }
 
-    public int calculatePrize(Integer lottoLankCount) {
+    public long calculatePrize(long lottoLankCount) {
         return getPrizeMoney() * lottoLankCount;
     }
